Update image files and add custom title bars
Updated several PNG image files including `circle_.png`, `CloseMacOS.png`, `close_.png`, `maximize-square-minimalistic-bold_.png`, `MaximizeMacOS.png`, `minimize-square-minimalistic-bold_.png`, `MinimizeMacOS.png`, and `window-minimize_.png`. The changes are represented by a series of characters, indicating that the content of these images has been updated or replaced.

Added a new `UserControl` in `MacosTitleBar.axaml` and `WindowsTitleBar.axaml` files that define the layout and appearance of a custom title bar for MacOS and Windows respectively. This includes the definition of the close, minimize, and zoom buttons, as well as the title and window icon.

Also added a new partial class `MacosTitleBar` and `WindowsTitleBar` in `MacosTitleBar.axaml.cs` and `WindowsTitleBar.axaml.cs` files respectively that handle the functionality of the custom title bar for MacOS and Windows. This includes methods for closing, minimizing, and maximizing the window, as well as subscribing to window state changes.