Added LoginScreenView and LoginScreenViewModel
Added a new class `LoginScreenViewModel` in the `LoginScreenViewModel.cs` file. This class is part of the `QuanLySinhVien.ViewModels.Login` namespace and inherits from `ViewModelBase`. Currently, this class is empty.

In the `LoginScreenView.axaml` file, a new Avalonia window was created with the class `QuanLySinhVien.Views.Login.LoginScreenView`. This window has a design width and height of 800 and 450 respectively, and its title is "LoginScreenView". The window currently only displays the text "Welcome to Avalonia!".

In the `LoginScreenView.axaml.cs` file, a new partial class `LoginScreenView` was created in the `QuanLySinhVien.Views.Login` namespace. This class inherits from `Window` and has a constructor that calls the `InitializeComponent` method. This class is likely to be used to handle events and interactions in the `LoginScreenView` window.