Updated LoginScreenView design and added DockPanel
Updated the design height and width of the LoginScreenView in the Avalonia XAML file from 450 and 800 to 550 and 900 respectively. Set the actual width and height of the window to 900 and 550. Removed the welcome message "Welcome to Avalonia!". Added a DockPanel to the window with a custom WindowsTitleBar with a height of 25. Added a Grid with two columns to the DockPanel, with a red Panel in the first column and a blue Panel in the second. In the C# code behind file, set the height of the TitleBar to -1 and the property ExtendsContentIntoTitleBar to true.