Updated UI elements and added new components
Updated the `FontFamily` and `WindowStartupLocation` of `LoginScreenView`. Changed the `Background` color of `WindowsTitleBar` and `Panel` in `Grid.Column="0"` and `Grid.Column="1"`. Added a new `StackPanel` to the `Panel` in `Grid.Column="1"` with specific dimensions and alignment. Inside this `StackPanel`, added a `Grid` with two `RowDefinition`s and two `TextBlock`s with specific text, alignment, and font properties. Also added an empty `StackPanel` to the `StackPanel` in `Grid.Column="1"`.