Updated UI control properties in LoginScreenView and WindowsTitleBar
Updated several properties in `LoginScreenView.axaml` and `WindowsTitleBar.axaml` to improve user interaction. The `IsTabStop` property was added to various controls to manage their accessibility via the Tab key. The `CanResize` property was added to the main window to prevent resizing. The `Padding` property of the TextBox style was adjusted for better content spacing. The `TabIndex` property was added to several controls to manage the order of access when using the Tab key.