Refactor MainWindow initialization and enhance navigation
Modified the `App` class in `App.axaml.cs` to change the `MainWindow` initialization process. The `MainWindow` is now assigned a new instance of `LoginScreenView` without a `DataContext` initially, which is then set to a new instance of `LoginScreenViewModel` with `MainWindow` as a parameter.

In `LoginScreenViewModel.cs`, added several namespaces and new fields to the `LoginScreenViewModel` class. A new constructor has been added that takes a `Window` parameter. The `OpenMainScreen` method has been modified to create a new instance of `MainScreenView`, set its `DataContext`, hide `_loginWindow`, and show `MainScreen`.

In `MainScreenView.axaml.cs`, the `MainScreenView` class now inherits from `AppWindow` instead of `Window`. A new constructor has been added that takes a `Window` parameter, sets the `TitleBar` properties, and adds an event handler to the `Closed` event.