Refactor LoginScreen and enhance MainScreen functionality
In this commit, the instantiation of `MainScreen` in `LoginScreenViewModel.cs` has been simplified by removing the setting of `DataContext`. In `MainScreenViewModel.cs`, new namespaces have been imported and two new private lists, `_viewModelList` and `_viewList`, have been added. A new public constructor and two new methods, `OnNavigateViewSelectionChanged` and `CreateInstance`, have been added. The `NavigationView` in `MainScreenView.axaml` has been named. In `MainScreenView.axaml.cs`, new namespaces have been imported, the `NavigationView` is found and stored in a variable, a new `MainScreenViewModel` is created, the `DataContext` of the view is set, and the `SelectionChanged` event is set to trigger `OnNavigateViewSelectionChanged`.