Updated navigation view and added new ViewResult view and view model
Updated `MainScreenViewModel.cs` to uncomment code for creating and setting instances of views and view models based on the selected navigation view item. Added a `Tag` attribute to a `NavigationViewItem` in `MainScreenView.axaml` for identifying the view and view model to be created when selected. Also added a new `NavigationViewItem` to the `FooterMenuItems`. Created new files `ViewResultViewModel.cs`, `ViewResultView.axaml`, and `ViewResultView.axaml.cs` to define a new view and view model, `ViewResultView` and `ViewResultViewModel`, and the code-behind for the new view.