Updated ViewResultViewModel and ViewResultView for real-time date/time
Updated the `ViewResultViewModel` class to include a new private string variable `_currentTime` and a public property `CurrentTime` that gets and sets `_currentTime` using the `RaiseAndSetIfChanged` method from the ReactiveUI library. Added a constructor for `ViewResultViewModel` that calls the `UpdateCurrentTime` method when a new instance of the class is created. The `UpdateCurrentTime` method is an asynchronous method that continuously updates the `CurrentTime` property every second with the current date and time in the format "HH:mm dd/MM/yyyy". In the `ViewResultView.axaml` file, the hardcoded text displaying the date and time has been replaced with a data binding to the `CurrentTime` property, allowing the displayed date and time to update in real time.