Added delete functionality to ViewResultView
Added a new property and command to `ViewResultViewModel.cs` to track and delete the selected row in the data grid. The `ViewResultView.axaml` file was updated to bind the delete button to the new `DeleteCommand`. In `ViewResultView.axaml.cs`, an event handler was added to update the `SelectedItemRow` property when the selected row changes.