"Refactor UI tags and adjust properties in XAML"
Refactored several UI tags in the XAML code for better readability and maintainability. The `DockPanel.Dock` and `ItemsSource` properties were moved to be self-closed. A commented out `StackPanel.Styles` section was added, including a `Style` for the `AutoCompleteBox`'s `TextBlock#PART_Watermark` selector with a `FontSize` of 5. The `Height` property was removed from both the `Button` and `AutoCompleteBox` tags, and the `Watermark` property was also removed from the `AutoCompleteBox` tag. The `IsTextCompletionEnabled` property was added to the `AutoCompleteBox` tag and set to `True`. Lastly, the closing `StackPanel` tag was moved to be self-closed.