Refactor ViewResultView UI and event handling
In this commit, the UI of `ViewResultView.axaml` has been significantly refactored. A `Grid` containing three `Button` elements was removed and replaced with a `StackPanel` containing `TextBox` and `TextBlock` elements, as well as two `DockPanel` elements. The `TextBox` elements have been simplified and styles have been added for `TextBlock` and `TextBox` elements.

In `ViewResultView.axaml.cs`, new `using` directives have been added and the `ViewResultView` constructor has been simplified. The `AttachHandler` method has been modified to include two new methods, `BindSearchText` and `BindDataGrid`.

`BindSearchText` sets up event handlers for the `KeyDown` and `TextChanged` events of a `TextBox` named `SearchTextBox`. `BindDataGrid` sets up event handlers for the `CellEditEnding` and `SelectionChanged` events of a `DataGrid`. The `BackupData` method of the `viewModel` is called when a cell's edit ends, and the selected row index and item properties are updated when the selection changes.