Updated UI elements and properties in axaml files
In this commit, several changes were made to the `Icons.axaml`, `TeacherInfoView.axaml`, and `ViewResultView.axaml` files. In `Icons.axaml`, two new `StreamGeometry` keys were added and the border brush color for several keys was updated. `TeacherInfoView.axaml` saw changes to the `Background`, `BorderBrush`, and `Margin` properties, the addition of a new `Grid`, and the removal of a button. In `ViewResultView.axaml`, the `Background` property was added and updated for the `Border` in the `Data Panel`. Lastly, the icons for buttons in both `TeacherInfoView.axaml` and `ViewResultView.axaml` were changed to use `PathIcon` with the respective `StreamGeometry` keys.