Added real-time clock and adjusted UI in TeacherInfoView
The `System.Globalization` namespace was added to the `TeacherInfoViewModel.cs` file to handle culture-related information. A new private string `_currentTime` and its public property `CurrentTime` were added to the same file, with the `CurrentTime` property using the `RaiseAndSetIfChanged` method from the `ReactiveUI` library to notify subscribers of changes. An asynchronous method `UpdateCurrentTime` was added to continuously update the `CurrentTime` property every second with the current date and time. This method is called in the constructor of the `TeacherInfoViewModel` class, starting the time update upon instance creation. In the `TeacherInfoView.axaml` file, hardcoded text was replaced with a binding to the `CurrentTime` property, displaying the updated current time. The width of the `DataGridTextColumn` that binds to `TeacherName` was adjusted from "2*" to "1.25*", reducing its relative space in the `DataGrid`.