Update UI elements in AddTeacherView and TeacherInfoView
In AddTeacherView.axaml, the FluentAvalonia.UI.Controls namespace was added under the alias 'ui'. The margin of a Grid element was adjusted, and several styles were added to a StackPanel to modify the appearance of TextBox and ComboBox elements. A new Grid was also added, containing an "Add Teacher" button with a green background and a "Cancel" button with a red background. In TeacherInfoView.axaml, the IsReadOnly property of a DataGrid was set to True.