Added ReactiveUI and student info features to StudentInfoViewModel
In `StudentInfoViewModel.cs`, ReactiveUI library was added for reactive programming. A new property `CurrentTime` was added to display the current time, along with an asynchronous method `UpdateCurrentTime` to update it every second. An `ObservableCollection` of `StudentInfoModel` named `ListStudentInfoModels` was added to store the list of students. The constructor was updated to initialize `ListStudentInfoModels` and add five `StudentInfoModel` objects to the list, and to call `UpdateCurrentTime` method.

In `StudentInfoView.axaml`, the view was updated with a white background, two new buttons for showing all students and adding a teacher, and adjusted margin of the search box. The `TextBlock` for displaying the current time was un-commented and bound to `CurrentTime` property. The old buttons for adding a teacher and showing all students were removed. The `DataGrid` was un-commented and updated to display student information instead of teacher information, with two new columns added for displaying the GPA of students for two semesters. The `DataGridTemplateColumn` was adjusted to modify the student information.

In `StudentInfoModel.cs`, a new model class `StudentInfoModel` was created with properties for storing student information such as name, birthday, gender, address, email, and GPA for two semesters.