**Summary:** Added functionality to open and close AddStudent window
**Details:**

In `AddStudentViewModel.cs`, a new dependency on `FluentAvalonia.UI.Windowing` was added and a private field `_addStudentWindow` of type `AppWindow` was introduced. Two constructors were added, one of which accepts an `AppWindow` object and assigns it to `_addStudentWindow`. A new method `CloseAddStudentWindow` was added to close the `_addStudentWindow`.

In `StudentInfoViewModel.cs`, a new dependency on `System.Reactive` and `QuanLySinhVien.Views.MainScreen` was added. A private field `addStudentCommand` of type `ReactiveCommand<Unit, Unit>` was introduced along with its public property `AddStudentCommand`. A new method `OpenAddStudentWindow` was added to create a new instance of `AddStudentView`, set its `DataContext` to a new instance of `AddStudentViewModel`, and show the view.

In `AddStudentView.axaml`, the margin and alignment of a `Grid` were changed. The `Command` property of a `Button` was bound to `CloseAddStudentWindow`.

In `AddStudentView.axaml.cs`, the `WindowStartupLocation` property of the window was set to `WindowStartupLocation.CenterOwner`.

In `StudentInfoView.axaml`, the `Command` property of a `Button` was bound to `OpenAddStudentWindow`.