Updated TeacherInfoViewModel and related UI components
In the `TeacherInfoViewModel.cs` file, the `addTeacherCommand` property has been replaced with an integer property `SelectedGiaoVienIndex` to store the index of the selected teacher. A new method `DeleteSelectedTeacher()` has been added to delete the selected teacher from the database and refresh the list.

In the `TeacherInfoView.axaml` file, the `DataGrid` control has been named for code-behind access, and a `Command` binding has been added to a `Button` to call the `DeleteSelectedTeacher()` method.

In the `TeacherInfoView.axaml.cs` file, a `SelectionChanged` event handler has been added to the `DataGrid` control to update the `SelectedGiaoVienIndex` property when the selected row changes. The `System.Diagnostics` and `QuanLySinhVien.ViewModels.MainScreen` namespaces have been added.