Added student management features in ViewModel and View
In this commit, we have made several enhancements to the student management system. In `AddStudentViewModel.cs`, new properties were added to store and manipulate student data, along with `AddCommand`, `OnAdd`, and `OnCancel` methods for student addition and window control. `StudentInfoViewModel.cs` now includes properties to store a list of students and methods to handle student addition and list loading. No changes were made to `TeacherInfoViewModel.cs`. In the View layer, `AddStudentView.axaml` and `AddStudentView.axaml.cs` were updated with bindings and event handlers to reflect changes in the ViewModel. Lastly, `StudentInfoView.axaml` was updated to pass the window as a parameter to the `OpenAddStudentWindow` command.