Added password visibility toggle feature in login screen
Implemented a new feature in the login screen that allows users to toggle the visibility of their password. This was achieved by adding a `RevealPassword` property and a `ChangePasswordTextBoxState` method in the `LoginScreenViewModel` class. A new button `RevealPasswordBtn` was added in the `LoginScreenView.axaml` file, which is bound to the `ChangePasswordTextBoxState` method. The `PasswordChar` property of the password TextBox is now bound to the `RevealPassword` property in the ViewModel. The icon of the `RevealPasswordBtn` button changes based on the value of `RevealPassword`. New `StreamGeometry` resources were added for various icons. The `ReactiveUI` namespace was imported for reactive programming in .NET.