Added Forgot Password functionality and UI updates
Added new `StreamGeometry` resources in `Icons.axaml` and a new method `OpenForgotPassword` in `LoginScreenViewModel.cs` to manage the Forgot Password functionality. Removed the `Background` attribute from the `StackPanel` in `LoginScreenView.axaml` and added a `Command` and `CommandParameter` to the `Button` for binding. Created `ForgotPasswordViewModel.cs` and `ForgotPasswordView.axaml` files for the new view and its operations. The `ForgotPasswordView` class, part of the `QuanLySinhVien.Views.Login` namespace, extends `AppWindow` and includes several new namespaces. The class is marked as partial and has a constructor for initialization and setting properties.