Add new properties and relationships to models
Added nullable properties `SoLuongDat` and `TiLe` to `BaoCaoMon`. Introduced navigation property `MaHocKyNavigation`. Enhanced `GiaoVien` with collections for `NguoiDungs` and `PhanCongGiangDays`. In `QlhsContext`, added `DbSet` properties for `NguoiDungs` and `NienKhoas`, and configured entity relationships and properties for `NguoiDung` and `NienKhoa`. Created `NguoiDung.cs` with key properties and a navigation link to `GiaoVien`.