Disable AddConfirm command in SubjectViewModel
The `AddConfirm` command within `SubjectViewModel.cs` has been commented out, disabling the functionality to add new subjects to the database through this command. This command was responsible for a series of checks and UI updates related to adding a subject, including validation of the subject name, user confirmation, uniqueness check in the database, and UI feedback upon adding a subject. With this change, the associated process for adding a subject via this command is no longer available.